# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "alwism.com - API Documentation",
        "description": "This is the API Gateway for [www.alwism.com](https://www.alwism.com).  You can find out more about alwism at [https://www.alwism.com/sitemechanism](https://www.alwism.com/sitemechanism) or visit to github repository [github](http://url.cc/).",
        "contact": {
            "name": "Contact the Developer",
            "url": "https://www.alwism.com/contactme"
        },
        "version": "1.0.3"
    },
    "servers": [
        {
            "url": "api/v1"
        }
    ],
    "paths": {
...
        "/api/admin/siteui/banner": {
            "post": {
                "tags": [
                    "siteUi"
                ],
                "summary": "Create or update the existing banner",
                "operationId": "e3f8cc0da0109b4ab8838266c61c49ab",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UiBanner"
                },
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/siteui/logo": {
            "post": {
                "tags": [
                    "siteUi"
                ],
                "summary": "Create or update the existing logo",
                "operationId": "b480af98744641c5cec5abeaa701b4fa",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UiLogo"
                },
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/siteui/footer": {
            "post": {
                "tags": [
                    "siteUi"
                ],
                "summary": "Create or update the existing footer",
                "operationId": "e091c267f51122b537ba353928e6c501",
                "requestBody": {
                    "$ref": "#/components/requestBodies/adfooter"
                },
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "requestBodies": {
...
            "UiBanner": {
                "description": "Ui Banner Request Body",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "title": "UiBanner",
                            "required": [
                                "image"
                            ],
                            "properties": {
                                "image": {
                                    "$ref": "#/components/schemas/SiteUi/properties/banner"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UiLogo": {
                "description": "Ui Banner Request Body",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "title": "UiBanner",
                            "required": [
                                "image"
                            ],
                            "properties": {
                                "image": {
                                    "$ref": "#/components/schemas/SiteUi/properties/logo"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "adfooter": {
                "description": "Footer Request Body",
                "required": true,
                "content": {
                    "multipart/x-www-form-urlencoded": {
                        "schema": {
                            "title": "footerText",
                            "required": [
                                "footerText",
                                "FcopyText"
                            ],
                            "properties": {
                                "footerText": {
                                    "$ref": "#/components/schemas/SiteUi/properties/footerText"
                                },
                                "FcopyText": {
                                    "$ref": "#/components/schemas/SiteUi/properties/FcopyText"
                                },
                                "FSkype": {
                                    "$ref": "#/components/schemas/SiteUi/properties/FSkype"
                                },
                                "FGithub": {
                                    "$ref": "#/components/schemas/SiteUi/properties/FGithub"
                                },
                                "FWhatsapp": {
                                    "$ref": "#/components/schemas/SiteUi/properties/FWhatsapp"
                                },
                                "Flinkedin": {
                                    "$ref": "#/components/schemas/SiteUi/properties/Flinkedin"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "application/json": {
                        "schema": {
                            "properties": {
                                "footerText": {
                                    "$ref": "#/components/schemas/SiteUi/properties/footerText"
                                },
                                "FcopyText": {
                                    "$ref": "#/components/schemas/SiteUi/properties/FcopyText"
                                },
                                "FSkype": {
                                    "$ref": "#/components/schemas/SiteUi/properties/FSkype"
                                },
                                "FGithub": {
                                    "$ref": "#/components/schemas/SiteUi/properties/FGithub"
                                },
                                "FWhatsapp": {
                                    "$ref": "#/components/schemas/SiteUi/properties/FWhatsapp"
                                },
                                "Flinkedin": {
                                    "$ref": "#/components/schemas/SiteUi/properties/Flinkedin"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}